backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.18
langchain==0.0.350
langchain-community==0.0.6
langchain-openai
sentence-transformers
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart==0.0.6
pydantic==2.5.2
jinja2==3.1.2
